add payments

+++ frontend/js/calcCartPrice.js
@@ -13,7 +13,7 @@ function calcCartPriceAndDelivery() {
 		// Находим количество товара
 		const amountEl = item.closest('.cart-item').querySelector('[data-counter]');
 		// Добавляем стоимость товара в общую стоимость (кол-во * цену)
-		priceTotal += parseInt(item.innerText) * parseInt(amountEl.innerText);
+		priceTotal += parseFloat(item.innerText) * parseInt(amountEl.innerText);
 	});
 
 	// Отображаем цену на странице
